requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
feedparser>=6.0.11
httpx==0.25.2
sqlalchemy>=2.0.25
aiosqlite==0.19.0
jinja2==3.1.2
python-dateutil==2.8.2
greenlet0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
feedparser==6.0.10
httpx==0.25.2
sqlalchemy>=2.0.25
aiosqlite==0.19.0
jinja2==3.1.2
python-dateutil==2.8.2
